csv: don't require NooBaa CRD
Unfortunately, since the NooBaa CSV requires the OB/OBC CRDs from us,
this creates a cyclic dependency... and in particular, OLM is terrible
about resolving this and thus leads to race conditions when installing
ODF 4.9. For now, we're removing the NooBaa CRD from the OCS CSV's
required CRDs. Since we don't support ever running ocs-operator outside
the context of ODF, this won't be an issue in production.

Signed-off-by: Jose A. Rivera <jarrpa@redhat.com>